// src/infrastructure/models/CanonicalCourseCSV.ts
export interface CanonicalCourseCSV {
  Periodo: string;
  Tipo: string;
  Asignatura: string;
  GRUPO: string;
  PE: string;
  Semestre: string;
  Horas_a_la_semana: string;
  Modalidad: string;
  Nombres: string;
  Apellidos: string;
  Creditos: string;
  Modelo: string;
  Lunes: string;
  Aula1: string;
  Martes: string;
  Aula2: string;
  Miercoles: string;
  Aula3: string;
  Jueves: string;
  Aula4: string;
  Viernes: string;
  Aula5: string;
}